cache draw points
